add README.md; Dockerfile and run the application on docker-compose; minor fixes

+++ src/test/java/io/group/ratelimiter/unit/infra/adapter/repository/QuotaRepositoryRedisAdapterUnitTest.java
@@ -33,7 +33,7 @@ class QuotaRepositoryRedisAdapterUnitTest extends AbstractUnitTest {
 
         var result = quotaRepositoryRedisAdapter.incrementQuota(userId);
 
-        assertEquals(incremented, result.getUsage());
+        assertEquals(incremented, result.getUsage().longValue());
     }
